refactor(feed): split the MT5 stream and the Python bridge under 1,500 lines
`crates/feed-mt5/src/stream.rs` owned the listener, both consumer ports, the
event vocabulary, the session loop, both block state machines, every publish
path and the line reader, at 2,200 production lines behind the last signed
ratchet entry in `feed-mt5`. `bridge/mt5/quantick_bridge.py` owned the whole
Python bridge at 1,981 lines, which the Rust guard cannot see at all.

`stream.rs` keeps the listener and the refusal path; six siblings own one
responsibility each: `ports.rs` the two shared-state handles, `events.rs` what
the server says, `connection.rs` the session loop, `blocks.rs` the two state
machines that outlive a line, `publish.rs` every send out, `reader.rs` the
bounded line reader.

`quantick_bridge.py` keeps its name, its CLI and the broker-clock cache, so
the configured `bridge_command` is unchanged. `Session` is now assembled from
four mixins beside it — transport, ticks, history and rates — over a shared
`quantick_bridge_core.py`.

Every body moved unchanged. The only Rust widenings are 23 `pub(super)` marks;
the Python `def`/`class` inventory matches item for item, and only the
`Session` class header changed to name its mixins. No behaviour, contract or
wire change, and no test expectation edited.

Three cross-surface agreement tests read a constant out of source by path, so
their paths follow their constants: `session_gap_agreement.rs` and the two
`include_str!` sites in `protocol.rs` now read
`bridge/mt5/quantick_bridge_core.py` and
`crates/feed-mt5/src/stream/connection.rs`. The agreements themselves are
unchanged. `bridge/mt5/tests/harness.py` gains `patch_bridge`, which fans a
module-global rebind out across every bridge module, because a single-module
rebind no longer reaches the code under test.

`crates/guards/size-baseline.txt`: the `crates/feed-mt5/src/stream.rs` entry is
retired (427 production lines, under the 1,500 threshold) and `!budget` falls
from 20,025 to 17,825, exactly the retired ceiling.
