+++ crates/feed-mt5/src/stream.rs
@@ -15,29 +15,37 @@
 //! Every noteworthy transition emits a structured `tracing` event with an
 //! `event_code` (see the diagnosis table in the crate docs, `lib.rs`): an AI
 //! or operator can reconstruct a session from logs alone.
+//!
+//! This file is the listener and the refusal path. The rest of a session is
+//! owned one module at a time beside it: [`ports`] the two handles the
+//! consumer holds, [`events`] what the server says, [`connection`] the
+//! session loop, [`blocks`] the two state machines that outlive a line,
+//! [`publish`] every send out, and [`reader`] the bounded line reader.
 
-use std::collections::BTreeMap;
 use std::net::SocketAddr;
-use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-use std::sync::{Arc, Mutex};
 use std::time::Duration;
-use std::time::{SystemTime, UNIX_EPOCH};
 
-use tokio::io::{AsyncBufReadExt as _, AsyncWriteExt as _, BufReader};
 use tokio::net::{TcpListener, TcpStream};
-use tokio::sync::{Notify, mpsc};
+use tokio::sync::mpsc;
 use tokio::task::JoinHandle;
-use tracing::{debug, info, warn};
+use tracing::{info, warn};
 
-use quantick_engine::{Bar, Trade};
-use quantick_orderbook::{DepthEvent, DepthResyncReason, DepthStatus};
+use crate::map::SideMode;
+use crate::protocol::{self, BridgeMsg};
 
-use crate::depth::BookMapper;
-use crate::latency::{LatencySample, LatencyTracker};
-use crate::map::{MapOutcome, PriceContext, SideMode, TickMapper};
-use crate::protocol::{self, BridgeMsg, FeedMsg, SCHEMA_VERSION, TapeKind};
-use crate::rates::RateMapper;
-use crate::session::SeqTracker;
+mod blocks;
+mod connection;
+mod events;
+mod ports;
+mod publish;
+mod reader;
+
+pub use events::{Mt5Error, Mt5Event, Mt5Status};
+pub use ports::{BookCaptureSwitch, HistoryPager};
+pub use reader::{BoundedLine, BoundedLineReader};
+
+use connection::serve_connection;
+use events::ConnEnd;
 
 /// Default address the feed listens on for the bridge.
 pub const DEFAULT_LISTEN_ADDR: &str = "127.0.0.1:9100";
@@ -74,206 +82,6 @@ const BUSY_REFUSAL_WINDOW: Duration = Duration::from_millis(250);
 /// looks. The two are meant to differ; if either moves, read the other first.
 const LAG_REPORT_MS: i64 = 1_000;
 
-/// Runtime switch controlling whether DOM images are published.
-///
-/// MT5's book shares one socket with ticks and MQL5 gives us no back-channel
-/// to ask the terminal to stop sending it, so "capture off" is a decision made
-/// here: images are decoded and dropped rather than published. That costs a
-/// JSON parse per image and nothing downstream — no book state, no history, no
-/// projection. The alternative (tearing down the bridge session) would take the
-/// trade stream down with it.
-///
-/// Cloning shares the switch; the consumer flips it from any thread.
-#[derive(Debug, Clone, Default)]
-pub struct BookCaptureSwitch(Arc<BookCaptureState>);
-
-#[derive(Debug, Default)]
-struct BookCaptureState {
-    enabled: AtomicBool,
-    /// Base generation chosen by the consumer. Each bridge session adds its own
-    /// offset on top, so a reconnect never reuses a generation.
-    base_generation: AtomicU64,
-}
-
-impl BookCaptureSwitch {
-    /// A switch that starts disabled.
-    #[must_use]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Publish depth from `base_generation` onwards.
-    ///
-    /// A base above the previous one discards whatever the old generation
-    /// published; consumers use that to keep stale in-flight events from an
-    /// earlier capture out of fresh history.
-    pub fn enable(&self, base_generation: u64) {
-        self.0
-            .base_generation
-            .store(base_generation, Ordering::Relaxed);
-        self.0.enabled.store(true, Ordering::Release);
-    }
-
-    /// Stop publishing depth. The bridge session and the trade stream continue.
-    pub fn disable(&self) {
-        self.0.enabled.store(false, Ordering::Release);
-    }
-
-    /// Whether depth is currently published.
-    #[must_use]
-    pub fn is_enabled(&self) -> bool {
-        self.0.enabled.load(Ordering::Acquire)
-    }
-
-    fn state(&self) -> (bool, u64) {
-        // Acquire on `enabled` pairs with the release in `enable`, so a true
-        // read never sees a stale base generation.
-        let enabled = self.0.enabled.load(Ordering::Acquire);
-        (enabled, self.0.base_generation.load(Ordering::Relaxed))
-    }
-}
-
-/// The consumer's one way to ask the terminal for ticks it has not sent — the
-/// back-channel behind the chart's "load older" button.
-///
-/// Shaped like [`BookCaptureSwitch`]: a shared handle the consumer holds and
-/// the running session reads, rather than a channel threaded through the
-/// server's signature. The two ports differ in what they carry — a switch has
-/// a state, a pager has a request — but not in how they dock, and a second
-/// shared-state port that invented its own plumbing would be the harder one to
-/// find.
-///
-/// **One request in flight per session.** A trader leaning on the button would
-/// otherwise queue pages the terminal answers minutes later, each landing in
-/// front of bars the earlier ones already drew. A request arriving while one is
-/// outstanding is dropped and counted, exactly as the UI already drops a click
-/// whose command channel is full.
-#[derive(Debug, Clone, Default)]
-pub struct HistoryPager(Arc<HistoryPagerState>);
-
-#[derive(Debug, Default)]
-struct HistoryPagerState {
-    /// Queue and gate under **one** lock.
-    ///
-    /// Not a mutex for the request and an atomic for the gate: the consumer
-    /// task and the session task run concurrently, and taking a request is
-    /// "clear the queue *and* raise the gate" — one decision. Split across two
-    /// primitives there is a window between them where a click reads a lowered
-    /// gate, queues, and gets sent while the previous page is still coming.
-    /// Locked once per click and once per session wake-up, never per tick.
-    request: Mutex<PagerRequest>,
-    /// Wakes the session loop, which is otherwise parked on the socket.
-    wake: Notify,
-}
-
-/// What the pager is holding: at most one queued request, and whether a fetch
-/// is already running.
-#[derive(Debug, Default)]
-struct PagerRequest {
-    /// The queued request, if any: `(count, before_utc_ms)`.
-    queued: Option<(u64, i64)>,
-    /// Raised while the session is fetching, so a second click is dropped
-    /// rather than queued behind the first.
-    in_flight: bool,
-}
-
-impl HistoryPager {
-    /// A pager with nothing pending.
-    #[must_use]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Ask for up to `count` ticks from before `before_utc_ms`.
-    ///
-    /// The cursor is **UTC** — the timestamp of the oldest [`Trade`] the
-    /// consumer holds, in the units it already has. The session converts it to
-    /// the terminal's clock on the way out, where the live offset lives.
-    ///
-    /// Returns `false` when the request was dropped because one is already in
-    /// flight — the caller's cue to leave its loading indicator alone rather
-    /// than start a second one.
-    pub fn request(&self, count: u64, before_utc_ms: i64) -> bool {
-        {
-            let mut held = self.0.request.lock().expect("history pager mutex");
-            // Queued counts as busy, not just in flight. The session task runs
-            // on another thread and may not have woken yet, so a click that
-            // only checked the gate would quietly overwrite the previous
-            // click's request — two asks, one answer, and a consumer counting
-            // replies left one short forever.
-            if held.in_flight || held.queued.is_some() {
-                return false;
-            }
-            held.queued = Some((count, before_utc_ms));
-        }
-        self.0.wake.notify_one();
-        true
-    }
-
-    /// Whether a request is outstanding.
-    #[must_use]
-    pub fn is_in_flight(&self) -> bool {
-        self.0
-            .request
-            .lock()
-            .expect("history pager mutex")
-            .in_flight
-    }
-
-    /// Park until a request arrives, then take it.
-    ///
-    /// Cancel-safe: [`Notify::notified`] and the take are separate steps, and a
-    /// notification that arrives before the wait does is remembered, so losing
-    /// this future to a `select!` cannot lose a click. Marks the request in
-    /// flight before returning it — the session is the only caller, and it is
-    /// about to write to the socket.
-    async fn take_request(&self) -> (u64, i64) {
-        loop {
-            {
-                let mut held = self.0.request.lock().expect("history pager mutex");
-                // Take and raise together. A click landing between the two
-                // would otherwise see a lowered gate and queue behind a page
-                // already on the wire.
-                if let Some(request) = held.queued.take() {
-                    held.in_flight = true;
-                    return request;
-                }
-            }
-            self.0.wake.notified().await;
-        }
-    }
-
-    /// The fetch is over (delivered, refused, or abandoned). Clears the gate so
-    /// the next click is heard, and reports whether a fetch was actually
-    /// running.
-    ///
-    /// The report is what tells a caller whether it still owes a reply. Callers
-    /// that already know they do can ignore it.
-    fn settle_owed(&self) -> bool {
-        let mut held = self.0.request.lock().expect("history pager mutex");
-        let owed = held.in_flight;
-        held.in_flight = false;
-        owed
-    }
-
-    /// Clear the pager at the end of a session, and say whether anything was
-    /// owed an answer.
-    ///
-    /// Both halves matter. The return value keeps the promise of one reply per
-    /// request — a request still *queued* when the socket died was never taken,
-    /// so the gate alone would miss it and leave a spinner running. And the
-    /// clearing is what keeps that promise from being kept twice: the caller
-    /// answers whatever this reports, so a request left behind would be served
-    /// again by the next session and arrive as a second reply to a click the
-    /// consumer already saw resolved.
-    fn abandon(&self) -> bool {
-        let mut held = self.0.request.lock().expect("history pager mutex");
-        let owed = held.queued.take().is_some() || held.in_flight;
-        held.in_flight = false;
-        owed
-    }
-}
-
 /// How the bridge server behaves for one symbol.
 #[derive(Debug, Clone)]
 pub struct ServerConfig {
@@ -309,193 +117,6 @@ impl ServerConfig {
             history_pager: HistoryPager::new(),
         }
     }
-}
-
-/// Where the feed currently stands, for honest labelling in UI and logs.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Mt5Status {
-    /// Listening; no bridge connected. The chart should say so, not pretend.
-    Waiting {
-        /// The actual bound address (resolves `:0` in tests).
-        addr: String,
-    },
-    /// A bridge said hello and is streaming (or about to).
-    Connected {
-        /// Symbol as configured.
-        symbol: String,
-        /// The front-month contract the terminal actually streams.
-        broker_symbol: String,
-        /// What this venue prints for the symbol, as the hello declared it.
-        ///
-        /// Only a live session knows: the same terminal streams an exchange
-        /// contract with a real tape and a broker CFD with none, and the
-        /// difference decides what the chart may honestly offer.
-        tape: TapeKind,
-        /// Levels per side this session can publish, or `None` when it sends no
-        /// depth at all (the terminal refused the DOM, or the symbol has none).
-        book_levels: Option<u32>,
-        /// Whether this session sends a historical candle block.
-        ///
-        /// Per-session like the two above: the Expert Advisor sends none, and
-        /// so does any bridge older than the feature. A consumer waiting on
-        /// candles needs to hear that now rather than after a timeout.
-        rates: bool,
-        /// Whether this session answers requests for older ticks.
-        ///
-        /// The one capability on this list the consumer can *act* on rather
-        /// than merely display: the chart's "load older" button is enabled by
-        /// this and by nothing else. Per-session for the same reason as the
-        /// rest — the same quantick build talks to a bridge that pages and to
-        /// one that does not, and the provider's name cannot tell them apart.
-        history_paging: bool,
-    },
-    /// The bridge went away; the server is looping back to waiting.
-    Lost {
-        /// Why, e.g. `"bye: deinit"`, `"silent"`, `"eof"`.
-        reason: String,
-    },
-}
-
-/// One message from the bridge server to its consumer.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Mt5Event {
-    /// A connection-state transition.
-    Status(Mt5Status),
-    /// One complete historical block (may be empty), already mapped. Sent
-    /// exactly once per `backfill_start`/`backfill_end` pair.
-    Backfilled(Vec<Trade>),
-    /// One live trade.
-    Live(Trade),
-    /// Where the tape's delay is being spent, measured at the socket.
-    ///
-    /// Sent at a bounded rate — at most once every
-    /// [`SAMPLE_EVERY_PRINTS`](crate::latency::SAMPLE_EVERY_PRINTS) prints, and
-    /// once per heartbeat so a thin tape still reports. Never once per print:
-    /// the reading costs a system clock read, and a per-print clock read on a
-    /// busy tape is the kind of cost this whole change exists to remove.
-    ///
-    /// The figures are taken where the line is read off the socket, so they
-    /// account for the chain up to this crate and no further. What a consumer
-    /// then spends queueing and drawing is its own to measure — and the
-    /// difference between its arrival figure and this one is exactly that.
-    Latency(LatencySample),
-    /// One order-book event, in the provider-neutral depth vocabulary.
-    ///
-    /// Only produced while [`BookCaptureSwitch`] is enabled and the bridge
-    /// declares depth support.
-    Depth(DepthEvent),
-    /// Another bridge dialed this port while a session was being served, and
-    /// was refused.
-    ///
-    /// The log has said this since the refusal existed; the chart has not, and
-    /// the chart is where someone is looking at a window that says "waiting for
-    /// the bridge" while the answer sits in a file. Carries the same diagnosis
-    /// the log gets, so the consumer can put it in front of a person.
-    SessionBusy {
-        /// Address of the connection that was turned away.
-        peer: String,
-        /// The symbol its hello declared, when it sent one.
-        peer_symbol: Option<String>,
-        /// Stable classification: `same_symbol`, `other_symbol`, or
-        /// `unidentified`.
-        diagnosis: &'static str,
-        /// What to do about it, in words.
-        advice: &'static str,
-    },
-    /// The session's historical candle block, complete and already mapped.
-    ///
-    /// Sent exactly once per `rates_start`/`rates_end` pair, ascending by
-    /// `open_time` and deduplicated. A block that never finished is discarded
-    /// rather than half-delivered — a candle series with a hole in the middle
-    /// reads as a market that stopped trading.
-    Rates {
-        /// Milliseconds each bar covers, as the block declared.
-        interval_ms: i64,
-        /// The candles, ascending by `open_time`.
-        bars: Vec<Bar>,
-        /// Whether the block is known to be short of what was asked for —
-        /// the bridge said so, or this decoder clipped it. See
-        /// [`protocol::BridgeMsg::RatesEnd`].
-        partial: bool,
-    },
-    /// The answer to one [`HistoryPager::request`]: ticks older than the cursor
-    /// the consumer asked from, already mapped and ascending by time.
-    ///
-    /// **Exactly one per request, always** — including when the terminal had
-    /// nothing, when the bridge cannot page, and when the session died before
-    /// answering. A consumer shows a spinner while it waits, and a request that
-    /// can go unanswered is a spinner that never stops.
-    HistoryPage {
-        /// The older trades, ascending. Empty is a legitimate answer.
-        trades: Vec<Trade>,
-        /// Whether the terminal reports nothing older left — the end of the
-        /// tape, not merely the end of this block. See
-        /// [`protocol::BridgeMsg::HistoryEnd`] for why an empty block alone
-        /// does not mean this.
-        exhausted: bool,
-        /// How far back the search actually reached, in **UTC** milliseconds,
-        /// when the bridge said.
-        ///
-        /// Distinct from the oldest trade in `trades`, and the difference is
-        /// what keeps paging moving over stretches that map to nothing — a
-        /// pre-open session of quote-only ticks, or a window that held none at
-        /// all. See [`protocol::BridgeMsg::HistoryEnd::scanned_to_ms`].
-        scanned_to_utc_ms: Option<i64>,
-    },
-    /// A slice of the *opening* history: older than everything sent so far, and
-    /// asked for by nobody.
-    ///
-    /// The bridge opens a chart on the whole trading session, which on a liquid
-    /// contract is far too much to arrive in one block. The newest slice goes
-    /// out as [`Mt5Event::Backfilled`] so there is something to look at within
-    /// a second, and the rest of the session follows in these, newest-first,
-    /// while the live tape keeps flowing between them.
-    ///
-    /// Deliberately **not** a [`Mt5Event::HistoryPage`]. That event's contract
-    /// is one reply per request and a consumer's spinner depends on it; these
-    /// answer no request and must settle no debt, or a click made during the
-    /// fill would be answered by a slice that has nothing to do with it.
-    OpeningPage {
-        /// The older trades, ascending. Empty is legitimate: a slice can map to
-        /// no trades at all.
-        trades: Vec<Trade>,
-        /// Slices still to come after this one, when the bridge said. For
-        /// showing progress; never a promise, since a session can end mid-fill.
-        remaining: Option<u64>,
-    },
-}
-
-/// A fatal server error (the non-fatal ones are events/logs).
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Mt5Error {
-    /// Could not bind the listen address (typically: port already in use by
-    /// another quantick instance).
-    Bind {
-        /// The address we tried.
-        addr: String,
-        /// The OS error text.
-        message: String,
-    },
-}
-
-impl std::fmt::Display for Mt5Error {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Mt5Error::Bind { addr, message } => {
-                write!(f, "cannot listen on {addr} for the MT5 bridge: {message}")
-            }
-        }
-    }
-}
-
-impl std::error::Error for Mt5Error {}
-
-/// Why one bridge connection ended.
-enum ConnEnd {
-    /// The consumer dropped the event channel: shut the server down.
-    UiGone,
-    /// The bridge went away (reason for the status event); keep listening.
-    BridgeGone(String),
 }
 
 /// Listen for the bridge and stream events until the consumer goes away.
@@ -795,1460 +416,12 @@ async fn identify<R: tokio::io::AsyncRead + Unpin>(source: R) -> PeerIdentity {
     }
 }
 
-/// Trades one paged block may hold before it stops growing.
-///
-/// The bridge is the side this crate cannot vouch for — the same reasoning
-/// behind [`MAX_BARS_PER_BLOCK`] for candles, and more pressing here: a candle
-/// block arrives once per session, a page can be opened again on every click.
-/// A peer that opens `history_start` and never stops would otherwise grow this
-/// vector until the feed task dies and takes the chart with it.
-///
-/// Sized well above any honest page: the consumer's own step tops out at 50 000
-/// (the toolbar's `DragValue` range) and the Python bridge caps itself at
-/// 200 000, so a block reaching this ceiling is a bridge that is not answering
-/// the question it was asked.
-const MAX_TRADES_PER_PAGE: usize = 250_000;
-
-/// A page of older ticks under construction, and the live tape's tick-rule
-/// context waiting for it to finish.
-///
-/// The context travels with the block rather than in a variable beside it so
-/// the two cannot be separated: every path that takes the block back also gets
-/// the context to put back, including the ones that discard it.
-struct PagedBlock {
-    /// The mapped trades collected so far.
-    trades: Vec<Trade>,
-    /// What the tick rule was reading before the block opened.
-    resume: PriceContext,
-    /// Trades dropped because the block hit [`MAX_TRADES_PER_PAGE`]. Reported
-    /// once at the end rather than per tick, so a runaway bridge cannot turn
-    /// the log into the second denial of service.
-    over_cap: u64,
-    /// Whether this block is a slice of the opening session rather than the
-    /// answer to a click. Carried from the start marker to the end one because
-    /// only the start says which kind it is, and only the end can act on it.
-    opening: bool,
-}
-
-/// What woke the session loop: something the bridge said, or something the
-/// trader asked for.
-///
-/// The two arrive on opposite halves of the same socket and the loop has to
-/// wait on both at once; naming them lets the wait stay one `select!` with one
-/// timeout rather than two loops racing over one reader.
-enum SessionInput {
-    /// A line from the bridge (or the error that ended the read).
-    Line(std::io::Result<BoundedLine>),
-    /// The consumer wants ticks older than `before_utc_ms`.
-    Request {
-        /// How many ticks to ask for.
-        count: u64,
-        /// Oldest UTC millisecond the consumer holds; the request is for
-        /// strictly older than this.
-        before_utc_ms: i64,
-    },
-}
-
-/// What happened to one attempt to ask the bridge for older ticks.
-enum PageRequestOutcome {
-    /// Written to the socket; the block will arrive on the read side.
-    Sent,
-    /// This bridge does not read its socket, so nothing was written.
-    Refused,
-    /// The write failed; the session is over.
-    WriteFailed(std::io::Error),
-}
-
-/// Put one request for older ticks on the wire.
-///
-/// Refuses rather than writes when the bridge never declared it reads: an
-/// unread request would sit in the peer's receive buffer, and on the Expert
-/// Advisor a full buffer eventually blocks the terminal thread that sends
-/// ticks — so a chart asking for history would stop the chart.
-async fn answer_page_request(
-    outgoing: &mut tokio::net::tcp::OwnedWriteHalf,
-    symbol: &str,
-    mapper: &TickMapper,
-    can_page: bool,
-    count: u64,
-    before_utc_ms: i64,
-) -> PageRequestOutcome {
-    if !can_page {
-        warn!(
-            target: "quantick::feed",
-            schema_version = 1_u8,
-            event_code = "MT5_LOAD_OLDER_UNSUPPORTED",
-            symbol,
-            requested = count,
-            action = "answer_empty",
-            advice = "update the bridge (bridge/mt5/quantick_bridge.py) to page older ticks",
-            "this bridge session does not answer requests for older ticks"
-        );
-        return PageRequestOutcome::Refused;
-    }
-    // The terminal stamps everything in server time, so the cursor crosses in
-    // its clock; the mapper owns that offset and its heartbeat refreshes.
-    let before_ms = mapper.to_server_ms(before_utc_ms);
-    let line = protocol::encode_line(&FeedMsg::LoadOlder { count, before_ms });
-    info!(
-        target: "quantick::feed",
-        schema_version = 1_u8,
-        event_code = "MT5_LOAD_OLDER_REQUESTED",
-        symbol,
-        requested = count,
-        before_ms,
-        before_utc_ms,
-        "asking the terminal for ticks older than the chart's oldest"
-    );
-    match outgoing.write_all(line.as_bytes()).await {
-        Ok(()) => PageRequestOutcome::Sent,
-        Err(e) => PageRequestOutcome::WriteFailed(e),
-    }
-}
-
-/// Serve one bridge connection to completion.
-///
-/// `generation_offset` is the server-wide capture-generation cursor; this
-/// function advances it whenever depth capture needs a fresh generation.
-async fn serve_connection(
-    stream: TcpStream,
-    config: &ServerConfig,
-    tx: &mpsc::Sender<Mt5Event>,
-    generation_offset: &mut u64,
-) -> ConnEnd {
-    // Split so the session can write while parked on a read. The write half is
-    // used at most once per trader click; the read half carries every tick.
-    let (incoming, mut outgoing) = stream.into_split();
-    let mut lines = BoundedLineReader::new(incoming);
-
-    // 1. The first message must be a hello that matches what we expect.
-    let hello = match tokio::time::timeout(config.hello_timeout, lines.next_line()).await {
-        Err(_) => {
-            warn!(
-                target: "quantick::feed",
-                schema_version = 1_u8,
-                event_code = "MT5_HELLO_TIMEOUT",
-                timeout_s = config.hello_timeout.as_secs(),
-                "connection said nothing; dropping it"
-            );
-            return ConnEnd::BridgeGone("hello timeout".to_string());
-        }
-        Ok(Err(e)) => {
-            warn!(
-                target: "quantick::feed",
-                schema_version = 1_u8,
-                event_code = "MT5_SOCKET_ERROR",
-                error = %e,
-                "socket error before hello; dropping the connection"
-            );
-            return ConnEnd::BridgeGone(format!("socket error before hello: {e}"));
-        }
-        Ok(Ok(BoundedLine::Eof)) => {
-            info!(
-                target: "quantick::feed",
-                schema_version = 1_u8,
-                event_code = "MT5_BRIDGE_EOF",
-                "connection closed before hello"
-            );
-            return ConnEnd::BridgeGone("closed before hello".to_string());
-        }
-        Ok(Ok(BoundedLine::TooLong)) => {
-            warn!(
-                target: "quantick::feed",
-                schema_version = 1_u8,
-                event_code = "MT5_LINE_TOO_LONG",
-                max_bytes = MAX_LINE_BYTES as u64,
-                "first line exceeded the size cap; dropping the connection"
-            );
-            return ConnEnd::BridgeGone("oversized hello".to_string());
-        }
-        Ok(Ok(BoundedLine::NotUtf8 { len })) => {
-            warn!(
-                target: "quantick::feed",
-                schema_version = 1_u8,
-                event_code = "MT5_UNDECODABLE_LINE",
-                error = "invalid utf-8",
-                line_bytes = len as u64,
-                "first line was not valid protocol; dropping the connection"
-            );
-            return ConnEnd::BridgeGone("undecodable hello".to_string());
-        }
-        Ok(Ok(BoundedLine::Line(line))) => match protocol::parse_line(&line) {
-            Ok(BridgeMsg::Hello(h)) => h,
-            Ok(other) => {
-                warn!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_PROTOCOL_VIOLATION",
-                    got = ?other,
-                    "first message was not a hello; dropping the connection"
-                );
-                return ConnEnd::BridgeGone("no hello".to_string());
-            }
-            Err(e) => {
-                warn!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_UNDECODABLE_LINE",
-                    error = %e,
-                    snippet = %snippet(&line),
-                    "first line was not valid protocol; dropping the connection"
-                );
-                return ConnEnd::BridgeGone("undecodable hello".to_string());
-            }
-        },
-    };
-
-    if hello.schema != SCHEMA_VERSION {
-        warn!(
-            target: "quantick::feed",
-            schema_version = 1_u8,
-            event_code = "MT5_SCHEMA_MISMATCH",
-            bridge_schema = hello.schema,
-            our_schema = SCHEMA_VERSION,
-            bridge = %hello.bridge,
-            bridge_version = %hello.bridge_version,
-            "bridge speaks a different protocol version; refusing"
-        );
-        return ConnEnd::BridgeGone(format!("schema mismatch (bridge {})", hello.schema));
-    }
-    if hello.symbol != config.symbol {
-        warn!(
-            target: "quantick::feed",
-            schema_version = 1_u8,
-            event_code = "MT5_SYMBOL_MISMATCH",
-            expected = %config.symbol,
-            got = %hello.symbol,
-            "bridge streams a different symbol than configured; refusing"
-        );
-        return ConnEnd::BridgeGone(format!("symbol mismatch ({})", hello.symbol));
-    }
-
-    info!(
-        target: "quantick::feed",
-        schema_version = 1_u8,
-        event_code = "MT5_HELLO_OK",
-        bridge = %hello.bridge,
-        bridge_version = %hello.bridge_version,
-        symbol = %hello.symbol,
-        broker_symbol = %hello.broker_symbol,
-        digits = hello.digits,
-        server_utc_offset_s = hello.server_utc_offset_s,
-        tape = ?hello.tape,
-        "bridge session established"
-    );
-    // Absent means no: writing to a bridge that never reads would fill its
-    // receive buffer and, on the EA, block the terminal thread that sends ticks.
-    let can_page = hello.history_paging.unwrap_or(false);
-    if tx
-        .send(Mt5Event::Status(Mt5Status::Connected {
-            symbol: hello.symbol.clone(),
-            broker_symbol: hello.broker_symbol.clone(),
-            tape: hello.tape,
-            book_levels: hello.book_levels,
-            rates: hello.rates.unwrap_or(false),
-            history_paging: can_page,
-        }))
-        .await
-        .is_err()
-    {
-        return ConnEnd::UiGone;
-    }
-    info!(
-        target: "quantick::feed",
-        schema_version = 1_u8,
-        event_code = if can_page {
-            "MT5_HISTORY_PAGING_AVAILABLE"
-        } else {
-            "MT5_HISTORY_PAGING_UNSUPPORTED"
-        },
-        symbol = %hello.symbol,
-        bridge = %hello.bridge,
-        bridge_version = %hello.bridge_version,
-        advice = if can_page {
-            "-"
-        } else {
-            "update the bridge (bridge/mt5/quantick_bridge.py) to page older ticks"
-        },
-        "whether this session can be asked for ticks older than it sent"
-    );
-
-    // 2. Stream messages until something ends the session.
-    //
-    // The bridge declares what its venue prints; the configured `side_mode` is
-    // a policy, `tape` is a fact about the instrument, and a fact the feed
-    // cannot observe for itself must come from the side that can see it.
-    let mut mapper =
-        TickMapper::new(config.side_mode, hello.server_utc_offset_s).with_tape(hello.tape);
-    let mut tracker = SeqTracker::new();
-    let mut latency = LatencyTracker::new();
-    // Whether the tape has already been reported late. Edge-triggered, so a
-    // session that stays behind logs the diagnosis once instead of once per
-    // sample, and its recovery is logged too — a report with no matching
-    // recovery is how an operator reads "it never came back".
-    let mut lag_reported = false;
-    // Whether the consumer's queue has already been reported full. Same edge
-    // trigger as the lag report, for the same reason.
-    let mut backpressure_reported = false;
-    let mut backfill: Option<Vec<Trade>> = None;
-    let mut candles: Option<RatesBlock> = None;
-    let mut undecodable: u64 = 0;
-    let mut depth = DepthSession::new(&hello, config.symbol.clone());
-    depth.log_capability();
-    // The paged block being collected, if the bridge is mid-answer, and the
-    // live tape's tick-rule context parked for the duration.
-    let mut page: Option<PagedBlock> = None;
-
-    // Parked here across the whole session rather than rebuilt each pass. The
-    // read loop polls this once per inbound line, and a future recreated every
-    // time would re-register with the `Notify` on every tick — paying a
-    // synchronization cost per print for something that fires when a trader
-    // clicks. Pinned once, a poll is a load and a return.
-    let pending_request = config.history_pager.take_request();
-    tokio::pin!(pending_request);
-
-    let end = loop {
-        // One wait covers both directions, under the bridge-liveness timeout.
-        //
-        // The timeout is rebuilt each pass, so a wake from *either* branch
-        // restarts it: a click buys a silent bridge one more `read_timeout`
-        // before it is declared lost. That is a bounded and rare extension — a
-        // click is a human action, and the pager allows one outstanding at a
-        // time — not an indefinite one, which is why the timeout stays out here
-        // rather than being tracked against the read alone.
-        let input = tokio::time::timeout(config.read_timeout, async {
-            tokio::select! {
-                // Biased so a busy tape cannot starve the click: this branch is
-                // ready at most once per trader action, the read branch on
-                // nearly every pass.
-                biased;
-                (count, before_utc_ms) = &mut pending_request => {
-                    pending_request.set(config.history_pager.take_request());
-                    SessionInput::Request { count, before_utc_ms }
-                }
-                // `next_line` is cancel-safe, and the reader outlives this
-                // `select!` so the state it keeps is still there next pass.
-                // Its awaits are all `fill_buf`, and every byte it takes from
-                // the `BufReader` is appended to the partial-line buffer in the
-                // same synchronous step that consumes it — including the
-                // multi-`fill_buf` path a line longer than the buffer takes.
-                // So a cancelled poll leaves the two exactly as consistent as
-                // an uncancelled one: nothing consumed is unrecorded, and
-                // nothing recorded is unconsumed.
-                line = lines.next_line() => SessionInput::Line(line),
-            }
-        })
-        .await;
-
-        let line = match input {
-            Err(_) => {
-                warn!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_BRIDGE_SILENT",
-                    timeout_s = config.read_timeout.as_secs(),
-                    "no ticks or heartbeats within the timeout; presuming the bridge dead"
-                );
-                break ConnEnd::BridgeGone("silent".to_string());
-            }
-            Ok(SessionInput::Request {
-                count,
-                before_utc_ms,
-            }) => {
-                match answer_page_request(
-                    &mut outgoing,
-                    &config.symbol,
-                    &mapper,
-                    can_page,
-                    count,
-                    before_utc_ms,
-                )
-                .await
-                {
-                    // Asked. The answer arrives as an ordinary block of lines
-                    // on the read side, like every other thing the bridge says.
-                    PageRequestOutcome::Sent => continue,
-                    // Nothing to ask: this bridge cannot page. The consumer is
-                    // still owed the one reply every request gets, or its
-                    // spinner runs forever.
-                    PageRequestOutcome::Refused => {
-                        config.history_pager.settle_owed();
-                        if tx
-                            .send(Mt5Event::HistoryPage {
-                                trades: Vec::new(),
-                                exhausted: false,
-                                scanned_to_utc_ms: None,
-                            })
-                            .await
-                            .is_err()
-                        {
-                            break ConnEnd::UiGone;
-                        }
-                        continue;
-                    }
-                    // A socket that cannot be written to cannot be read from
-                    // either; end the session and let the reconnect handle it.
-                    // The owed reply is sent by the tail below, which covers
-                    // every way a session can end while a page is outstanding.
-                    PageRequestOutcome::WriteFailed(e) => {
-                        break ConnEnd::BridgeGone(format!("socket error on write: {e}"));
-                    }
-                }
-            }
-            Ok(SessionInput::Line(Err(e))) => {
-                warn!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_SOCKET_ERROR",
-                    error = %e,
-                    "socket error; dropping the session"
-                );
-                break ConnEnd::BridgeGone(format!("socket error: {e}"));
-            }
-            Ok(SessionInput::Line(Ok(BoundedLine::Eof))) => {
-                info!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_BRIDGE_EOF",
-                    "bridge closed the socket"
-                );
-                break ConnEnd::BridgeGone("eof".to_string());
-            }
-            Ok(SessionInput::Line(Ok(BoundedLine::TooLong))) => {
-                warn!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_LINE_TOO_LONG",
-                    max_bytes = MAX_LINE_BYTES as u64,
-                    "peer streamed an oversized line; dropping the session"
-                );
-                break ConnEnd::BridgeGone("oversized line".to_string());
-            }
-            Ok(SessionInput::Line(Ok(BoundedLine::NotUtf8 { len }))) => {
-                undecodable += 1;
-                warn!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_UNDECODABLE_LINE",
-                    error = "invalid utf-8",
-                    line_bytes = len as u64,
-                    total_undecodable = undecodable,
-                    "skipping an undecodable line"
-                );
-                continue;
-            }
-            Ok(SessionInput::Line(Ok(BoundedLine::Line(line)))) => line,
-        };
-        if line.trim().is_empty() {
-            continue;
-        }
-
-        match protocol::parse_line(&line) {
-            Err(e) => {
-                undecodable += 1;
-                warn!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_UNDECODABLE_LINE",
-                    error = %e,
-                    snippet = %snippet(&line),
-                    total_undecodable = undecodable,
-                    "skipping an undecodable line"
-                );
-            }
-            Ok(BridgeMsg::Tick(tick)) => {
-                let _ = tracker.observe(tick.seq);
-                if let MapOutcome::Trade { trade, .. } = mapper.map(&tick) {
-                    // A tick belongs to whichever block is open around it. The
-                    // paged block is checked first because it is the one that
-                    // can open mid-session: everything else is live by default,
-                    // and a paged tick delivered as live would append history to
-                    // the front of the tape.
-                    match (page.as_mut(), backfill.as_mut()) {
-                        (Some(block), _) => {
-                            if block.trades.len() < MAX_TRADES_PER_PAGE {
-                                block.trades.push(trade);
-                            } else {
-                                block.over_cap = block.over_cap.saturating_add(1);
-                            }
-                        }
-                        (None, Some(buf)) => buf.push(trade),
-                        (None, None) => {
-                            // Live prints only. A backfill or paged tick is as
-                            // old as the history it belongs to, and measuring
-                            // latency from one would report minutes of delay on
-                            // a chart that is perfectly current.
-                            latency.observe_live(tick.time_ms, tick.sent_ms);
-                            // Sampled *before* the trade is handed downstream,
-                            // never after. `send_live` waits when the consumer
-                            // is not draining, and a clock read on the far side
-                            // of that wait charges the consumer's own queueing
-                            // to the wire — collapsing the very gap the health
-                            // view publishes these two figures to expose.
-                            if latency.due()
-                                && publish_latency(
-                                    &mut latency,
-                                    &mapper,
-                                    &config.symbol,
-                                    &mut lag_reported,
-                                    tx,
-                                )
-                                .await
-                                .is_err()
-                            {
-                                break ConnEnd::UiGone;
-                            }
-                            if send_live(tx, trade, &config.symbol, &mut backpressure_reported)
-                                .await
-                                .is_err()
-                            {
-                                break ConnEnd::UiGone;
-                            }
-                        }
-                    }
-                }
-            }
-            Ok(BridgeMsg::Book(image)) => {
-                match depth
-                    .observe(image, &config.book_capture, generation_offset, tx)
-                    .await
-                {
-                    Ok(()) => {}
-                    Err(()) => break ConnEnd::UiGone,
-                }
-            }
-            Ok(BridgeMsg::Heartbeat(hb)) => {
-                if let Some(offset) = hb.server_utc_offset_s {
-                    mapper.set_server_utc_offset_s(offset);
-                    depth.set_server_utc_offset_s(offset);
-                }
-                // The beat a thin tape is measured on: a symbol printing once
-                // a minute never reaches the per-print sampling bound, and
-                // would otherwise never report at all. The figures are still
-                // the newest print's own, so a quiet stretch reports how late
-                // that print was and never how long ago it was.
-                if publish_latency(&mut latency, &mapper, &config.symbol, &mut lag_reported, tx)
-                    .await
-                    .is_err()
-                {
-                    break ConnEnd::UiGone;
-                }
-                // A heartbeat is the natural moment to notice that a consumer
-                // is waiting for depth this bridge cannot send.
-                if depth
-                    .report_missing_capability(&config.book_capture, tx)
-                    .await
-                    .is_err()
-                {
-                    break ConnEnd::UiGone;
-                }
-                debug!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_HEARTBEAT",
-                    seq_last = hb.seq_last,
-                    ticks_sent = hb.ticks_sent,
-                    "bridge heartbeat"
-                );
-            }
-            Ok(BridgeMsg::BackfillStart { count_hint }) => {
-                info!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_BACKFILL_START",
-                    count_hint = ?count_hint,
-                    "bridge is sending history"
-                );
-                backfill = Some(Vec::new());
-            }
-            Ok(BridgeMsg::BackfillEnd {}) => {
-                let batch = backfill.take().unwrap_or_default();
-                info!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_BACKFILL_END",
-                    trades = batch.len(),
-                    "history block complete"
-                );
-                if tx.send(Mt5Event::Backfilled(batch)).await.is_err() {
-                    break ConnEnd::UiGone;
-                }
-            }
-            Ok(BridgeMsg::HistoryStart {
-                count_hint,
-                opening,
-            }) => {
-                if !opening && !config.history_pager.is_in_flight() {
-                    // Nobody asked. Collect it anyway rather than letting its
-                    // ticks fall through as live prints — the block is history,
-                    // and charting it at the front of the tape is the one
-                    // outcome worse than dropping it. Logged under the same
-                    // code as the matching end: one condition, one thing to
-                    // grep for, both halves of the event findable.
-                    warn!(
-                        target: "quantick::feed",
-                        schema_version = 1_u8,
-                        event_code = "MT5_HISTORY_PAGE_UNSOLICITED",
-                        action = "collect_and_discard",
-                        "history_start with no request outstanding"
-                    );
-                }
-                info!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_HISTORY_PAGE_START",
-                    count_hint = ?count_hint,
-                    "bridge is sending a page of older ticks"
-                );
-                // A block already open means the bridge restarted mid-send.
-                // Put the live context back before parking it again, or the
-                // second start would park an already-emptied one and the tape
-                // would never get its price back.
-                if let Some(stale) = page.take() {
-                    mapper.restore_price_context(stale.resume);
-                }
-                // The page's ticks are hours older than the live tape; the tick
-                // rule must read them against each other, not against the
-                // newest price on the chart.
-                page = Some(PagedBlock {
-                    trades: Vec::new(),
-                    resume: mapper.take_price_context(),
-                    over_cap: 0,
-                    opening,
-                });
-            }
-            Ok(BridgeMsg::HistoryEnd {
-                exhausted,
-                scanned_to_ms,
-                remaining,
-            }) => {
-                let Some(PagedBlock {
-                    trades,
-                    resume,
-                    over_cap,
-                    opening,
-                }) = page.take()
-                else {
-                    // The start went missing — an undecodable line ahead of it
-                    // is enough. The block's ticks have already been charted as
-                    // live, which cannot be undone here; what must not also
-                    // happen is the pager staying latched. Without settling,
-                    // `request` refuses every later click for the rest of the
-                    // session and the button silently stops working.
-                    warn!(
-                        target: "quantick::feed",
-                        schema_version = 1_u8,
-                        event_code = "MT5_PROTOCOL_VIOLATION",
-                        action = "settle_and_answer_empty",
-                        "history_end without a history_start; its ticks went out as live"
-                    );
-                    if config.history_pager.settle_owed()
-                        && tx
-                            .send(Mt5Event::HistoryPage {
-                                trades: Vec::new(),
-                                exhausted: false,
-                                scanned_to_utc_ms: None,
-                            })
-                            .await
-                            .is_err()
-                    {
-                        break ConnEnd::UiGone;
-                    }
-                    continue;
-                };
-                // The live tape resumes from the price it left off at, never
-                // from wherever the page ended.
-                mapper.restore_price_context(resume);
-                if over_cap > 0 {
-                    warn!(
-                        target: "quantick::feed",
-                        schema_version = 1_u8,
-                        event_code = "MT5_HISTORY_PAGE_TRUNCATED",
-                        cap = MAX_TRADES_PER_PAGE as u64,
-                        dropped = over_cap,
-                        "a page exceeded the per-block cap; the surplus was dropped"
-                    );
-                }
-                if opening {
-                    // Nobody asked for this, so nothing is settled: a click
-                    // made while the session is still filling in stays owed its
-                    // own reply. The slice is still prepended — it is the
-                    // trader's morning.
-                    info!(
-                        target: "quantick::feed",
-                        schema_version = 1_u8,
-                        event_code = "MT5_OPENING_PAGE",
-                        trades = trades.len(),
-                        remaining = ?remaining,
-                        "a slice of the opening session is ready to prepend"
-                    );
-                    if tx
-                        .send(Mt5Event::OpeningPage { trades, remaining })
-                        .await
-                        .is_err()
-                    {
-                        break ConnEnd::UiGone;
-                    }
-                    continue;
-                }
-                if !config.history_pager.settle_owed() {
-                    warn!(
-                        target: "quantick::feed",
-                        schema_version = 1_u8,
-                        event_code = "MT5_HISTORY_PAGE_UNSOLICITED",
-                        trades = trades.len(),
-                        action = "discard",
-                        "a page nobody asked for; discarding it rather than prepending"
-                    );
-                    continue;
-                }
-                info!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_HISTORY_PAGE_END",
-                    trades = trades.len(),
-                    exhausted,
-                    scanned_to_ms = ?scanned_to_ms,
-                    "page of older ticks complete"
-                );
-                // Converted here, where the offset lives, for the same reason
-                // the outbound cursor is: the consumer speaks UTC and this is
-                // the one place that tracks what the terminal's clock is doing.
-                let scanned_to_utc_ms = scanned_to_ms.map(|ms| mapper.to_utc_ms(ms));
-                if tx
-                    .send(Mt5Event::HistoryPage {
-                        trades,
-                        exhausted,
-                        scanned_to_utc_ms,
-                    })
-                    .await
-                    .is_err()
-                {
-                    break ConnEnd::UiGone;
-                }
-            }
-            Ok(BridgeMsg::RatesStart {
-                interval_ms,
-                count_hint,
-            }) => {
-                info!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_RATES_START",
-                    interval_ms,
-                    count_hint = ?count_hint,
-                    "bridge is sending historical candles"
-                );
-                if let Some(open) = candles.take() {
-                    // A block already running: the bridge restarted mid-send,
-                    // or two are interleaved. Either way what was collected
-                    // belongs to a window this new header does not describe.
-                    warn!(
-                        target: "quantick::feed",
-                        schema_version = 1_u8,
-                        event_code = "MT5_PROTOCOL_VIOLATION",
-                        bars = open.len(),
-                        action = "discard_open_block",
-                        "a second rates_start arrived inside an open block; discarding the first"
-                    );
-                }
-                candles = Some(RatesBlock::new(interval_ms, hello.server_utc_offset_s));
-            }
-            Ok(BridgeMsg::Rate(chunk)) => match candles.as_mut() {
-                Some(block) => block.absorb(&chunk),
-                // A batch outside a block has no interval to be measured in,
-                // and guessing one would misdate every bar in it.
-                None => warn!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_PROTOCOL_VIOLATION",
-                    bars = chunk.bars.len(),
-                    action = "drop_chunk",
-                    "candles arrived outside a rates block; dropping them"
-                ),
-            },
-            Ok(BridgeMsg::RatesEnd { partial }) => match candles.take() {
-                Some(block) => {
-                    let (interval_ms, bars, clipped) = block.finish(&config.symbol);
-                    // Either side may know the block is short: the bridge from
-                    // its paging, this decoder from its own cap. Kept apart in
-                    // the log — they point at different things to go fix.
-                    let bridge_said_partial = partial;
-                    let partial = bridge_said_partial || clipped;
-                    if partial {
-                        warn!(
-                            target: "quantick::feed",
-                            schema_version = 1_u8,
-                            event_code = "MT5_RATES_PARTIAL",
-                            symbol = %config.symbol,
-                            bars = bars.len(),
-                            bridge_said_partial,
-                            clipped_here = clipped,
-                            "the candle block is short of what was asked for"
-                        );
-                    }
-                    if tx
-                        .send(Mt5Event::Rates {
-                            interval_ms,
-                            bars,
-                            partial,
-                        })
-                        .await
-                        .is_err()
-                    {
-                        break ConnEnd::UiGone;
-                    }
-                }
-                None => warn!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_PROTOCOL_VIOLATION",
-                    action = "ignore",
-                    "rates_end without a rates_start; ignoring it"
-                ),
-            },
-            Ok(BridgeMsg::Bye { reason }) => {
-                info!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_BRIDGE_BYE",
-                    reason = %reason,
-                    "bridge said goodbye"
-                );
-                break ConnEnd::BridgeGone(format!("bye: {reason}"));
-            }
-            Ok(BridgeMsg::Hello(_)) => {
-                warn!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_PROTOCOL_VIOLATION",
-                    "second hello mid-session; ignoring it"
-                );
-            }
-        }
-    };
-
-    if backfill.is_some() {
-        warn!(
-            target: "quantick::feed",
-            schema_version = 1_u8,
-            event_code = "MT5_PARTIAL_BACKFILL_DISCARDED",
-            "session ended mid-backfill; discarding the incomplete block"
-        );
-    }
-    // Every request gets exactly one reply, including the ones this session
-    // died holding — whether it had been sent to the bridge or was still
-    // queued. The alternative is a spinner that outlives the connection it was
-    // waiting on. `abandon` also clears the request itself: it belongs to the
-    // connection that is ending, and replaying it against the next session
-    // would page from a cursor that one never sent.
-    if config.history_pager.abandon() {
-        warn!(
-            target: "quantick::feed",
-            schema_version = 1_u8,
-            event_code = "MT5_HISTORY_PAGE_UNANSWERED",
-            partial_trades = page.as_ref().map_or(0, |block| block.trades.len()),
-            action = "answer_empty",
-            "session ended with a page outstanding; answering it empty"
-        );
-        let _ = tx
-            .send(Mt5Event::HistoryPage {
-                trades: Vec::new(),
-                exhausted: false,
-                scanned_to_utc_ms: None,
-            })
-            .await;
-    }
-    if let Some(block) = candles {
-        warn!(
-            target: "quantick::feed",
-            schema_version = 1_u8,
-            event_code = "MT5_PARTIAL_RATES_DISCARDED",
-            bars = block.len(),
-            "session ended mid-candle-block; discarding it (the next session re-sends)"
-        );
-    }
-    mapper.stats.log_summary(&config.symbol);
-    // A consumer that was capturing depth must hear that this generation ended,
-    // so it renders the discontinuity instead of connecting liquidity across it.
-    depth.close(tx).await;
-    end
-}
-
-/// Most candles one block may deliver.
-///
-/// The bridge caps what it sends (`--rates-max-bars`) and logs the shortfall,
-/// but the bridge is the side this one cannot vouch for: a misconfigured or
-/// hostile one could stream candles until the feed runs out of memory. Ninety
-/// days of one-minute buckets is ~130 000, so this is comfortably above any
-/// legitimate block while still being a bound.
-const MAX_BARS_PER_BLOCK: usize = 1_000_000;
-
-/// The historical candle block being received, between `rates_start` and
-/// `rates_end`.
-///
-/// Bars land in a [`BTreeMap`] keyed by `open_time` rather than a `Vec`: the
-/// terminal can repeat a bucket across chunk boundaries, and a map both settles
-/// that (last write wins — a repeat is a correction) and hands back one
-/// ascending series without a sort whose tie-breaking would be an unstated
-/// rule. Same block in, same series out, whatever order the chunks arrived in.
-struct RatesBlock {
-    interval_ms: i64,
-    mapper: RateMapper,
-    bars: BTreeMap<i64, Bar>,
-    /// Whether the cap has been hit, so it is reported once rather than per row.
-    truncated: bool,
-}
-
-impl RatesBlock {
-    fn new(interval_ms: i64, server_utc_offset_s: i64) -> Self {
-        Self {
-            interval_ms,
-            mapper: RateMapper::new(interval_ms, server_utc_offset_s),
-            bars: BTreeMap::new(),
-            truncated: false,
-        }
-    }
-
-    /// Map and absorb one chunk. Unreadable rows are counted, not fatal: one
-    /// corrupt candle in ninety days is a gap, not a reason to lose the block.
-    fn absorb(&mut self, chunk: &protocol::RateChunk) {
-        for row in &chunk.bars {
-            if self.bars.len() >= MAX_BARS_PER_BLOCK {
-                if !self.truncated {
-                    self.truncated = true;
-                    warn!(
-                        target: "quantick::feed",
-                        schema_version = 1_u8,
-                        event_code = "MT5_RATES_TRUNCATED",
-                        max_bars = MAX_BARS_PER_BLOCK as u64,
-                        action = "keep_oldest_stop_absorbing",
-                        "the candle block exceeded the cap; ignoring the rest of it"
-                    );
-                }
-                return;
-            }
-            if let Some(bar) = self.mapper.map(row) {
-                self.bars.insert(bar.open_time, bar);
-            }
-        }
-    }
-
-    fn len(&self) -> usize {
-        self.bars.len()
-    }
-
-    /// Close the block: log what it cost, and hand back the ascending series.
-    fn finish(self, symbol: &str) -> (i64, Vec<Bar>, bool) {
-        self.mapper.stats.log_summary(symbol, self.interval_ms);
-        // Clipping here is the same kind of shortfall the bridge reports with
-        // its own `partial`: bars that exist and were not delivered.
-        let clipped = self.truncated;
-        (self.interval_ms, self.bars.into_values().collect(), clipped)
-    }
-}
-
-/// Depth capture state for one bridge connection.
-///
-/// Split out because it is the only stateful thing in the message loop besides
-/// tick mapping, and it must stay correct across three independent events: the
-/// consumer toggling capture, the terminal losing images, and the session
-/// ending.
-struct DepthSession {
-    symbol: String,
-    /// `None` when the bridge declared no Depth of Market support.
-    mapper: Option<BookMapper>,
-    /// Whether the consumer has been told a generation is open.
-    publishing: bool,
-    last_seq: Option<u64>,
-    missing_capability_reported: bool,
-}
-
-impl DepthSession {
-    fn new(hello: &protocol::Hello, symbol: String) -> Self {
-        Self {
-            mapper: hello.book_levels.map(|levels| {
-                BookMapper::new(
-                    symbol.clone(),
-                    0,
-                    Some(levels),
-                    hello.tick_size.as_deref(),
-                    hello.server_utc_offset_s,
-                )
-            }),
-            symbol,
-            publishing: false,
-            last_seq: None,
-            missing_capability_reported: false,
-        }
-    }
-
-    fn log_capability(&self) {
-        match &self.mapper {
-            Some(_) => info!(
-                target: "quantick::feed",
-                schema_version = 1_u8,
-                event_code = "MT5_BOOK_AVAILABLE",
-                symbol = %self.symbol,
-                "bridge declares Depth of Market support"
-            ),
-            None => info!(
-                target: "quantick::feed",
-                schema_version = 1_u8,
-                event_code = "MT5_BOOK_UNSUPPORTED_BY_BRIDGE",
-                symbol = %self.symbol,
-                action = "trades_only",
-                "bridge declares no Depth of Market; the heatmap will stay empty \
-                 (recompile bridge/mt5/QuantickBridge.mq5, or the terminal refused the DOM)"
-            ),
-        }
-    }
-
-    fn set_server_utc_offset_s(&mut self, offset_s: i64) {
-        if let Some(mapper) = self.mapper.as_mut() {
-            mapper.set_server_utc_offset_s(offset_s);
-        }
-    }
-
-    /// Handle one image. `Err(())` means the consumer is gone.
-    async fn observe(
-        &mut self,
-        image: protocol::Book,
-        capture: &BookCaptureSwitch,
-        generation_offset: &mut u64,
-        tx: &mpsc::Sender<Mt5Event>,
-    ) -> Result<(), ()> {
-        if self.mapper.is_none() {
-            // A bridge sending images it never declared is a version skew, not
-            // data to trust silently.
-            return Ok(());
-        }
-        let (enabled, base_generation) = capture.state();
-        let lost_images = self.images_lost(image.seq);
-        let mapper = self.mapper.as_mut().expect("checked above");
-        if !enabled {
-            if self.publishing {
-                let generation = mapper.generation();
-                self.publishing = false;
-                self.last_seq = None;
-                mapper.restart(generation); // next capture starts from a snapshot
-                send_depth_status(tx, &self.symbol, generation, DepthStatus::Stopped).await?;
-            }
-            return Ok(());
-        }
-
-        // Open a generation when capture starts, when the consumer moves its
-        // base, or when images were lost and the diff would silently bridge a
-        // moment we never observed.
-        let wanted = base_generation.saturating_add(*generation_offset);
-        if !self.publishing || mapper.generation() != wanted || lost_images {
-            if lost_images {
-                send_depth_status(
-                    tx,
-                    &self.symbol,
-                    mapper.generation(),
-                    DepthStatus::Resyncing {
-                        reason: DepthResyncReason::SourceRestarted {
-                            cause: "book_images_lost",
-                        },
-                    },
-                )
-                .await?;
-            }
-            *generation_offset = generation_offset.saturating_add(1);
-            let generation = base_generation.saturating_add(*generation_offset);
-            mapper.restart(generation);
-            self.publishing = true;
-            send_depth_status(tx, &self.symbol, generation, DepthStatus::Connecting).await?;
-        }
-        self.last_seq = Some(image.seq);
-
-        let Some(event) = mapper.map(&image) else {
-            return Ok(());
-        };
-        let synchronized =
-            matches!(event, DepthEvent::Snapshot { .. }).then(|| mapper.synchronized_status());
-        let generation = mapper.generation();
-        if tx.send(Mt5Event::Depth(event)).await.is_err() {
-            return Err(());
-        }
-        if let Some(status) = synchronized {
-            send_depth_status(tx, &self.symbol, generation, status).await?;
-        }
-        Ok(())
-    }
-
-    /// Whether images went missing (or the bridge restarted its counter)
-    /// between the last one and `seq`.
-    fn images_lost(&self, seq: u64) -> bool {
-        match self.last_seq {
-            Some(last) => seq != last.saturating_add(1),
-            None => false,
-        }
-    }
-
-    /// Tell a waiting consumer, once, that this bridge cannot supply depth.
-    async fn report_missing_capability(
-        &mut self,
-        capture: &BookCaptureSwitch,
-        tx: &mpsc::Sender<Mt5Event>,
-    ) -> Result<(), ()> {
-        let (enabled, base_generation) = capture.state();
-        if self.mapper.is_some() || self.missing_capability_reported || !enabled {
-            return Ok(());
-        }
-        self.missing_capability_reported = true;
-        warn!(
-            target: "quantick::feed",
-            schema_version = 1_u8,
-            event_code = "MT5_BOOK_UNSUPPORTED_BY_BRIDGE",
-            symbol = %self.symbol,
-            action = "report_disconnected",
-            "depth capture is on but this bridge sends no Depth of Market"
-        );
-        // Tagged with the consumer's own base generation: a status below the
-        // generation floor it is watching would be discarded as stale, and the
-        // chart would keep waiting for a book that is never coming.
-        send_depth_status(
-            tx,
-            &self.symbol,
-            base_generation,
-            DepthStatus::Disconnected {
-                error_class: "bridge_without_depth",
-            },
-        )
-        .await
-    }
-
-    /// End the generation when the bridge session ends.
-    async fn close(&mut self, tx: &mpsc::Sender<Mt5Event>) {
-        if let Some(mapper) = self.mapper.as_ref() {
-            mapper.stats.log_summary(&self.symbol);
-            if self.publishing {
-                let _ = send_depth_status(
-                    tx,
-                    &self.symbol,
-                    mapper.generation(),
-                    DepthStatus::Disconnected {
-                        error_class: "bridge_lost",
-                    },
-                )
-                .await;
-            }
-        }
-    }
-}
-
-/// Publish one live trade, saying so when the consumer's queue is full.
-///
-/// `Err(())` means the consumer is gone.
-///
-/// A full queue here is not a harmless wait. This task is the one reading the
-/// socket, so a blocked publish stops the read, the receive buffer fills, and
-/// the bridge's next `SocketSend` blocks *the terminal's main thread* — the
-/// thread that would otherwise be collecting ticks. One slow consumer therefore
-/// becomes a late tape at the source, and until now it did that in silence:
-/// every hop looked healthy and the chart simply ran behind.
-///
-/// `try_send` first costs nothing on the happy path and is the only way to tell
-/// "the queue was full" from "the send took a while", so the wait is still
-/// taken — never a dropped print — but it is a wait with a name on it.
-async fn send_live(
-    tx: &mpsc::Sender<Mt5Event>,
-    trade: Trade,
-    symbol: &str,
-    reported: &mut bool,
-) -> Result<(), ()> {
-    match tx.try_send(Mt5Event::Live(trade)) {
-        Ok(()) => {
-            if *reported {
-                *reported = false;
-                info!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_CONSUMER_KEEPING_UP",
-                    symbol = %symbol,
-                    "the consumer's queue drained; the socket is being read freely again"
-                );
-            }
-            Ok(())
-        }
-        Err(mpsc::error::TrySendError::Full(event)) => {
-            if !*reported {
-                *reported = true;
-                warn!(
-                    target: "quantick::feed",
-                    schema_version = 1_u8,
-                    event_code = "MT5_CONSUMER_BACKPRESSURE",
-                    symbol = %symbol,
-                    action = "wait_for_room",
-                    "the consumer is not draining live trades fast enough; \
-                     this stops the socket read, which stalls the bridge's sends"
-                );
-            }
-            tx.send(event).await.map_err(|_| ())
-        }
-        Err(mpsc::error::TrySendError::Closed(_)) => Err(()),
-    }
-}
-
-/// Draw a latency sample, log the diagnosis when it crosses, and publish it.
-///
-/// `Err(())` means the consumer is gone, the same signal every other publish in
-/// this module returns.
-///
-/// This is the one place in the read path that reads a system clock. It is
-/// called at most once every
-/// [`SAMPLE_EVERY_PRINTS`](crate::latency::SAMPLE_EVERY_PRINTS) prints and once
-/// per heartbeat, so a tape printing a thousand times a second pays for
-/// roughly sixteen clock reads to be measurable — not a thousand.
-async fn publish_latency(
-    latency: &mut LatencyTracker,
-    mapper: &TickMapper,
-    symbol: &str,
-    lag_reported: &mut bool,
-    tx: &mpsc::Sender<Mt5Event>,
-) -> Result<(), ()> {
-    let Some(sample) = latency.sample(wall_clock_ms(), mapper.server_utc_offset_ms()) else {
-        // No live print since the last sample: nothing happened to measure, and
-        // republishing the previous window would let a wedged socket show a
-        // healthy split forever.
-        return Ok(());
-    };
-    let late = sample.arrival_lag_ms >= LAG_REPORT_MS;
-    if late != *lag_reported {
-        *lag_reported = late;
-        // Named, not absorbed. A tape that falls behind used to be visible only
-        // as a number drifting up in the corner of the chart, with nothing
-        // anywhere saying which hop was spending the time.
-        if late {
-            warn!(
-                target: "quantick::feed",
-                schema_version = 1_u8,
-                event_code = "MT5_TAPE_LATE",
-                symbol = %symbol,
-                arrival_lag_ms = sample.arrival_lag_ms,
-                terminal_lag_ms = sample.terminal_lag_ms,
-                terminal_lag_peak_ms = sample.terminal_lag_peak_ms,
-                transport_lag_ms = sample.transport_lag_ms,
-                prints = sample.prints,
-                hop = sample.dominant().map(crate::latency::LatencyHop::label),
-                "the tape is running behind; the hop field says where the time went"
-            );
-        } else {
-            info!(
-                target: "quantick::feed",
-                schema_version = 1_u8,
-                event_code = "MT5_TAPE_CAUGHT_UP",
-                symbol = %symbol,
-                arrival_lag_ms = sample.arrival_lag_ms,
-                "the tape is current again"
-            );
-        }
-    }
-    tx.send(Mt5Event::Latency(sample)).await.map_err(|_| ())
-}
-
-/// UTC epoch milliseconds.
-///
-/// The only wall-clock read in this crate, and it exists so a delay can be
-/// named. Everything that turns bridge lines into trades stays a pure function
-/// of what arrived.
-fn wall_clock_ms() -> i64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map(|d| i64::try_from(d.as_millis()).unwrap_or(i64::MAX))
-        .unwrap_or(0)
-}
-
-/// Publish one depth status. `Err(())` means the consumer is gone.
-async fn send_depth_status(
-    tx: &mpsc::Sender<Mt5Event>,
-    symbol: &str,
-    generation: u64,
-    status: DepthStatus,
-) -> Result<(), ()> {
-    tx.send(Mt5Event::Depth(DepthEvent::Status {
-        symbol: symbol.to_string(),
-        generation,
-        status,
-    }))
-    .await
-    .map_err(|_| ())
-}
-
 /// First 120 chars of a line, for log context without flooding. Truncates on
 /// a char boundary: a byte-index slice would panic mid-codepoint.
 fn snippet(line: &str) -> &str {
     match line.char_indices().nth(120) {
         Some((i, _)) => &line[..i],
         None => line,
-    }
-}
-
-#[cfg(test)]
-mod bounded_reader_tests {
-    use super::{BoundedLine, BoundedLineReader, MAX_LINE_BYTES};
-
-    /// The reader is generic so the app can point it at a launched bridge's
-    /// stderr, not just a socket. A cursor stands in for both.
-    #[tokio::test]
-    async fn it_reads_lines_from_any_source_not_just_a_socket() {
-        let source = std::io::Cursor::new(b"first\nsecond\r\n".to_vec());
-        let mut reader = BoundedLineReader::new(source);
-        assert_eq!(
-            reader.next_line().await.unwrap(),
-            BoundedLine::Line("first".to_owned())
-        );
-        // A CRLF terminator loses the carriage return, like the socket path.
-        assert_eq!(
-            reader.next_line().await.unwrap(),
-            BoundedLine::Line("second".to_owned())
-        );
-        assert_eq!(reader.next_line().await.unwrap(), BoundedLine::Eof);
-    }
-
-    #[tokio::test]
-    async fn an_endless_line_is_capped_and_reading_continues() {
-        // The failure this bound exists for: output with no newline in it.
-        // The oversized line is dropped, and the line after it still arrives.
-        let mut bytes = vec![b'x'; MAX_LINE_BYTES + 10];
-        bytes.push(b'\n');
-        bytes.extend_from_slice(b"survivor\n");
-        let mut reader = BoundedLineReader::new(std::io::Cursor::new(bytes));
-        assert_eq!(reader.next_line().await.unwrap(), BoundedLine::TooLong);
-        assert_eq!(
-            reader.next_line().await.unwrap(),
-            BoundedLine::Line("survivor".to_owned())
-        );
-    }
-
-    #[tokio::test]
-    async fn invalid_utf8_is_reported_and_skipped_not_fatal() {
-        // A Windows-encoded path inside a Python traceback is exactly this.
-        let mut bytes = b"before\n".to_vec();
-        bytes.extend_from_slice(&[0xFF, 0xFE, b'\n']);
-        bytes.extend_from_slice(b"after\n");
-        let mut reader = BoundedLineReader::new(std::io::Cursor::new(bytes));
-        assert_eq!(
-            reader.next_line().await.unwrap(),
-            BoundedLine::Line("before".to_owned())
-        );
-        assert_eq!(
-            reader.next_line().await.unwrap(),
-            BoundedLine::NotUtf8 { len: 2 }
-        );
-        assert_eq!(
-            reader.next_line().await.unwrap(),
-            BoundedLine::Line("after".to_owned()),
-            "one unreadable line must not end the stream"
-        );
-    }
-}
-
-/// One read from a [`BoundedLineReader`].
-#[derive(Debug, PartialEq, Eq)]
-pub enum BoundedLine {
-    /// A complete UTF-8 line (terminator stripped).
-    Line(String),
-    /// A complete line that was not valid UTF-8; skippable, per PROTOCOL.md.
-    NotUtf8 {
-        /// Length of the rejected line, in bytes.
-        len: usize,
-    },
-    /// The peer streamed more than [`MAX_LINE_BYTES`] without a newline.
-    TooLong,
-    /// The peer closed the connection.
-    Eof,
-}
-
-/// What one `fill_buf` round decided (split out so the borrow of the reader's
-/// internal buffer ends before `consume`).
-enum ReadStep {
-    Eof,
-    Line(usize),
-    TooLong(usize),
-    More(usize),
-}
-
-/// A newline-delimited reader that never buffers more than
-/// [`MAX_LINE_BYTES`], unlike `AsyncBufReadExt::lines`. Cancel-safe: the only
-/// await is `fill_buf`, and bytes move out of the source buffer and into the
-/// line buffer within a single poll.
-///
-/// Generic over the source because the bridge speaks the same line-delimited
-/// shape over two transports: a socket, and the stderr of a bridge quantick
-/// launched itself. Both need the same bound — an unterminated line is a
-/// memory leak wherever it comes from — and one implementation is what keeps
-/// the two honest about it.
-pub struct BoundedLineReader<R> {
-    reader: BufReader<R>,
-    buf: Vec<u8>,
-}
-
-impl<R: tokio::io::AsyncRead + Unpin> BoundedLineReader<R> {
-    /// Wrap `source`, reading at most [`MAX_LINE_BYTES`] per line.
-    pub fn new(source: R) -> Self {
-        Self {
-            reader: BufReader::new(source),
-            buf: Vec::new(),
-        }
-    }
-
-    /// The next line, or what went wrong with it.
-    pub async fn next_line(&mut self) -> std::io::Result<BoundedLine> {
-        loop {
-            let step = {
-                let available = self.reader.fill_buf().await?;
-                if available.is_empty() {
-                    ReadStep::Eof
-                } else if let Some(pos) = available.iter().position(|&b| b == b'\n') {
-                    if self.buf.len() + pos > MAX_LINE_BYTES {
-                        ReadStep::TooLong(pos + 1)
-                    } else {
-                        self.buf.extend_from_slice(&available[..pos]);
-                        ReadStep::Line(pos + 1)
-                    }
-                } else if self.buf.len() + available.len() > MAX_LINE_BYTES {
-                    ReadStep::TooLong(available.len())
-                } else {
-                    self.buf.extend_from_slice(available);
-                    ReadStep::More(available.len())
-                }
-            };
-            match step {
-                ReadStep::Eof => {
-                    // A trailing unterminated line still counts, like
-                    // `lines()` behaves.
-                    if self.buf.is_empty() {
-                        return Ok(BoundedLine::Eof);
-                    }
-                    return Ok(Self::finish(std::mem::take(&mut self.buf)));
-                }
-                ReadStep::Line(consume) => {
-                    self.reader.consume(consume);
-                    return Ok(Self::finish(std::mem::take(&mut self.buf)));
-                }
-                ReadStep::TooLong(consume) => {
-                    self.reader.consume(consume);
-                    self.buf.clear();
-                    return Ok(BoundedLine::TooLong);
-                }
-                ReadStep::More(consume) => self.reader.consume(consume),
-            }
-        }
-    }
-
-    fn finish(mut line: Vec<u8>) -> BoundedLine {
-        if line.last() == Some(&b'\r') {
-            line.pop();
-        }
-        match String::from_utf8(line) {
-            Ok(text) => BoundedLine::Line(text),
-            Err(e) => BoundedLine::NotUtf8 {
-                len: e.as_bytes().len(),
-            },
-        }
     }
 }
