+++ crates/guards/size-baseline.txt
@@ -240,7 +240,8 @@
 # `drain_feed` opening-block note with it: that arm is in `tab/feed.rs` and
 # was never what this entry paid for. Nothing was deleted; the budget falls
 # by the three ceilings.
-!budget 14673
+# +2: the `QUANTICK_FOOTPRINT_DEBUG` declaration. See the budget note above.
+!budget 12473
 
 crates/app/src/app.rs 769
 
@@ -353,11 +354,6 @@ crates/app/src/app.rs 769
 # `the_report_numbers_are_fixed` unchanged on both sides of it.
 
 crates/app/src/control/gateway.rs 1696
-# Raised for the same branch: an `Mt5Event::OpeningPage` variant and one
-# branch in the block state machine deciding whether a block settles a
-# request. The session loop is the only place that knows what a
-# `history_start` opened; the slicing itself lives in the bridge.
-crates/feed-mt5/src/stream.rs 2200
 crates/orderflow/src/projection.rs 1850
 crates/app/src/control/contract.rs 1718
 crates/app/src/control/evidence.rs 1708
